src/main/java/ru/hogwarts/school/service/implementation/FacultyServiceImpl.java
package ru.hogwarts.school.service.implementation;

import org.springframework.stereotype.Service;
import ru.hogwarts.school.model.Faculty;
import ru.hogwarts.school.repository.FacultyRepository;
import ru.hogwarts.school.service.interfaces.FacultyService;


import java.util.List;

@Service
public class FacultyServiceImpl implements FacultyService {

    private final FacultyRepository facultyRepository;

    public FacultyServiceImpl(FacultyRepository facultyRepository) {
        this.facultyRepository = facultyRepository;
    }


    @Override
    public Faculty create(Faculty faculty) {
        return facultyRepository.save(faculty);
    }

    @Override
    public Faculty reade(long id) {
        return facultyRepository.findById(id).get();
    }

    @Override
    public Faculty update(Faculty faculty) {
        return facultyRepository.save(faculty);
    }


    @Override
    public void delete(long id) {
        facultyRepository.deleteById(id);
    }

    @Override
    public List<Faculty> findByColor(String color) {
        return facultyRepository.findByColor(color);

    }
}